refactor. info in src/refactor.txt
