+++ src/refactor.txt
@@ -1,0 +1,33 @@
+lib.rs:
+    ALL macro_rules! definitions go here
+    run_app()
+
+coordinate.rs:
+    anything to do with the coordinate/grid structure
+    Coordinate
+    Row
+    Col
+
+grammar.rs:
+    anything to do with grammars
+    Grammar
+    Kind
+    Interactive
+
+model.rs:
+    anything to do with the state of the application
+     still almost 800 lines, maybe could be split up more
+    Model
+    SideMenu
+    Action
+    view_ functions
+
+session.rs:
+    all saving/loading code will go here
+    Session
+
+style.rs:
+    Style
+
+utils.rs:
+    all misc/helper function that AREN'T macros go here
