+++ src/style.rs
@@ -1,0 +1,47 @@
+use std::option::Option;
+use serde::Deserialize;
+
+use crate::coordinate::Coordinate;
+use crate::grammar::{Grammar, Kind};
+use crate::model::Model;
+
+// Style contains the relevant CSS properties for styling
+// a grammar Cell or Grid
+#[derive(Deserialize, Debug, Clone)]
+pub struct Style {
+    pub width: f64,  // CSS: width
+    pub height: f64, // CSS: height
+    pub border_color: String,  // CSS: border-color
+    pub border_collapse: bool, // CSS: border-collapse
+    pub font_weight: i32,      // CSS: font-weight
+    pub font_color: String,    // CSS: font-color
+}
+js_serializable!( Style );
+js_deserializable!( Style );
+
+impl Style {
+    pub fn default() -> Style {
+        Style {
+            width: 90.00,
+            height: 30.00,
+            border_color: "grey".to_string(),
+            border_collapse: false,
+            font_weight: 400,
+            font_color: "black".to_string(),
+        }
+    }
+
+    pub fn to_string(&self) -> String {
+        format!{
+        "/* border: 1px; NOTE: ignoring Style::border_* for now */
+border-collapse: {};
+font-weight: {};
+color: {};\n",
+        // self.border_color,
+        if self.border_collapse { "collapse" } else { "inherit" },
+        self.font_weight,
+        self.font_color,
+        }
+    }
+}
+
